+++ bin/differential_hash_expression.py
@@ -21,31 +21,41 @@ from tqdm import tqdm
 import sourmash_utils
 
 MAX_GROUP_SIZE = 100
-GROUP = 'group'
-SIG = 'sig'
-FASTA = 'fasta'
+GROUP = "group"
+SIG = "sig"
+FASTA = "fasta"
 
 
 # Default backend for scikit-learn logistic regression
-PENALTY = 'l1'
-SOLVER = 'saga'
+PENALTY = "l1"
+SOLVER = "saga"
 
+MIN_ABUNDANCE = 2
+MIN_CELLS = 3
+COEF_COL = "coefficient"
 
 # Create a logger
-logging.basicConfig(format='%(name)s - %(asctime)s %(levelname)s: %(message)s')
+logging.basicConfig(format="%(name)s - %(asctime)s %(levelname)s: %(message)s")
 logger = logging.getLogger(__file__)
 logger.setLevel(logging.INFO)
 
 
 def make_hash_df(sigs, with_abundance=False):
     if with_abundance:
-        records = {x.name(): x.minhash.get_mins(with_abundance=with_abundance)
-                   for x in sigs}
+        records = {
+            x.name(): x.minhash.get_mins(with_abundance=with_abundance) for x in sigs
+        }
     else:
         # Set value of each hash abundance to 1
-        records = {x.name(): dict.fromkeys(x.minhash.get_mins(with_abundance=with_abundance), 1)
-                   for x in sigs}
-    return pd.DataFrame(records)
+        records = {
+            x.name(): dict.fromkeys(
+                x.minhash.get_mins(with_abundance=with_abundance), 1
+            )
+            for x in sigs
+        }
+    hash_df = pd.DataFrame(records)
+
+    return hash_df
 
 
 def make_target_vector(n_group1, n_group2):
@@ -54,23 +64,43 @@ def make_target_vector(n_group1, n_group2):
     return y_target
 
 
-def get_training_data(sigs1, sigs2, with_abundance=False, verbose=False):
+def get_training_data(
+    sigs1,
+    sigs2,
+    with_abundance=False,
+    verbose=False,
+    min_cells=MIN_CELLS,
+    min_abundance=MIN_ABUNDANCE,
+):
     """Create X feature matrix and y target vector for machine learning"""
 
     # Create pandas dataframe of hash abundances
-    hash_df1 = make_hash_df(sigs1, with_abundance=with_abundance)
+    hash_df1 = make_hash_df(
+        sigs1,
+        with_abundance=with_abundance
+    )
     logger.info(f"Group1 hash dataframe head: {hash_df1.head()}")
 
-    hash_df2 = make_hash_df(sigs2, with_abundance=with_abundance)
+    hash_df2 = make_hash_df(
+        sigs2,
+        with_abundance=with_abundance
+    )
     logger.info(f"Group2 hash dataframe head: {hash_df2.head()}")
 
-    logger.info(f'Number of hashes in group1: {len(hash_df1.index)}')
-    logger.info(f'Number of hashes in group2: {len(hash_df2.index)}')
+    logger.info(f"Number of hashes in group1: {len(hash_df1.index)}")
+    logger.info(f"Number of hashes in group2: {len(hash_df2.index)}")
 
     # Concatenate to make feature matrix
     hash_df = pd.concat([hash_df1, hash_df2], axis=1)
     X = hash_df.T
     X = X.fillna(0)
+
+    if with_abundance and min_cells:
+        # Filter for hashes with at least min_abundance abundance, in at least
+        # min_cells samples
+        hash_mask = X >= min_abundance
+        hash_filter = hash_mask.sum() > min_cells
+        X = X.loc[:, hash_filter]
 
     # Create target vector "group1" is 1s and everything else is 0
     y_target = make_target_vector(len(sigs1), len(sigs2))
@@ -78,31 +108,60 @@ def get_training_data(sigs1, sigs2, with_abundance=False, verbose=False):
     return X, y_target
 
 
-def differential_hash_expression(sigs1, sigs2, with_abundance=False, verbose=False,
-                                 penalty=PENALTY, solver=SOLVER,
-                                 random_state=0, class_weight='balanced',
-                                 # Smaller C for stronger regularization
-                                 # (fewer final features to look at) --> only the good stuff is left
-                                 # This also (seems to) help with convergence?
-                                 C=0.1,
-                                 **kwargs):
+def differential_hash_expression(
+    sigs1,
+    sigs2,
+    with_abundance=False,
+    min_abundance=MIN_ABUNDANCE,
+    min_cells=MIN_CELLS,
+    verbose=False,
+    penalty=PENALTY,
+    solver=SOLVER,
+    random_state=0,
+    class_weight="balanced",
+    # Smaller C for stronger regularization
+    # (fewer final features to look at) --> only the good stuff is left
+    # This also (seems to) help with convergence?
+    C=0.1,
+    **kwargs,
+):
     if verbose:
         print("Creating training data")
-    X, y = get_training_data(sigs1, sigs2, with_abundance=with_abundance,
-                             verbose=verbose)
+    X, y = get_training_data(
+        sigs1,
+        sigs2,
+        with_abundance=with_abundance,
+        verbose=verbose,
+        min_cells=min_cells,
+        min_abundance=min_abundance,
+    )
 
-    regressor = LogisticRegression(solver=solver, penalty=penalty, verbose=verbose,
-                                   random_state=random_state, class_weight=class_weight,
-                                   C=C, **kwargs)
+    regressor = LogisticRegression(
+        solver=solver,
+        penalty=penalty,
+        verbose=verbose,
+        random_state=random_state,
+        class_weight=class_weight,
+        C=C,
+        **kwargs,
+    )
     logger.info(f"Running logistic regression: {regressor}")
     regressor.fit(X, y)
 
-    coefficients = pd.Series(regressor.coef_[0], index=X.columns)
-    n_positive = (coefficients > regressor.tol).sum()
-    logger.info(f'Number of coefficients greater than tolerance '
-                f'(tolerance: {regressor.tol}): {n_positive}')
+    coefficients = pd.Series(regressor.coef_[0], index=X.columns, name=COEF_COL)
 
-    return coefficients
+    # Sum the hash abundance per group
+    sums = X.groupby(y, axis=0).sum().astype(int)
+    sums = sums.T
+    coeffs_medians = sums.join(coefficients)
+
+    n_positive = (coefficients > regressor.tol).sum()
+    logger.info(
+        f"Number of coefficients greater than tolerance "
+        f"(tolerance: {regressor.tol}): {n_positive}"
+    )
+
+    return coeffs_medians
 
 
 def maybe_subsample(sigs, subsample_groups=MAX_GROUP_SIZE, random_state=0):
@@ -113,9 +172,19 @@ def maybe_subsample(sigs, subsample_groups=MAX_GROUP_SIZE, random_state=0):
     return sigs
 
 
-def get_hashes_enriched_in_group(group1_name, annotations, group_col, sketch_series,
-                                 max_group_size=MAX_GROUP_SIZE, random_state=0,
-                                 verbose=False, with_abundance=False, **kwargs):
+def get_hashes_enriched_in_group(
+    group1_name,
+    annotations,
+    group_col,
+    sketch_series,
+    max_group_size=MAX_GROUP_SIZE,
+    random_state=0,
+    verbose=False,
+    with_abundance=False,
+    min_cells=MIN_CELLS,
+    min_abundance=MIN_ABUNDANCE,
+    **kwargs,
+):
     rows = annotations[group_col] == group1_name
 
     group1_samples = annotations.loc[rows].index.intersection(sketch_series.index)
@@ -123,27 +192,50 @@ def get_hashes_enriched_in_group(group1_name, annotations, group_col, sketch_ser
 
     # Everything not in group 1
     group2_samples = annotations.loc[~rows].index.intersection(sketch_series.index)
-    logger.info(f"\nNumber of samples in the rest -- aka NOT {group1_name}: {len(group1_samples)}")
+    logger.info(
+        f"\nNumber of samples in the rest -- aka NOT {group1_name}: {len(group1_samples)}"
+    )
 
     group1_sigs = maybe_subsample(sketch_series[group1_samples], max_group_size)
     group2_sigs = maybe_subsample(sketch_series[group2_samples], max_group_size)
-    logger.info(f'\nGroup 1 signatures: {group1_sigs}')
-    logger.info(f'\nGroup 2 signatures: {group2_sigs}')
+    logger.info(f"\nGroup 1 signatures: {group1_sigs}")
+    logger.info(f"\nGroup 2 signatures: {group2_sigs}")
 
-    coefficients = differential_hash_expression(group1_sigs, group2_sigs,
-                                                verbose=verbose,
-                                                random_state=random_state,
-                                                with_abundance=with_abundance,
-                                                **kwargs)
-    coefficients.name = group1_name
+    coefficients = differential_hash_expression(
+        group1_sigs,
+        group2_sigs,
+        verbose=verbose,
+        random_state=random_state,
+        with_abundance=with_abundance,
+        min_cells=min_cells,
+        min_abundance=min_abundance,
+        **kwargs,
+    )
+    coefficients = coefficients.rename(columns={0: group1_name, 1: "rest"})
     return coefficients
 
 
-def main(metadata_csv, ksize, molecule, group_col=GROUP, group1=None, sig_col=SIG,
-         threshold=0, verbose=True, C=0.1, solver=SOLVER, penalty=PENALTY, n_jobs=8,
-         random_state=0, use_sig_basename=False, with_abundance=False,
-          max_group_size=MAX_GROUP_SIZE):
-    metadata = pd.read_csv(metadata_csv, index_col='sample_id')
+def main(
+    metadata_csv,
+    ksize,
+    molecule,
+    group_col=GROUP,
+    group1=None,
+    sig_col=SIG,
+    threshold=0,
+    verbose=True,
+    C=0.1,
+    solver=SOLVER,
+    penalty=PENALTY,
+    n_jobs=8,
+    random_state=0,
+    use_sig_basename=False,
+    with_abundance=False,
+    min_cells=3,
+    min_abundance=MIN_ABUNDANCE,
+    max_group_size=MAX_GROUP_SIZE,
+):
+    metadata = pd.read_csv(metadata_csv, index_col="sample_id")
 
     if use_sig_basename:
         metadata[sig_col] = metadata[sig_col].map(os.path.basename)
@@ -154,84 +246,127 @@ def main(metadata_csv, ksize, molecule, group_col=GROUP, group1=None, sig_col=SI
     logger.info(f"\nLoaded {len(sketches)} sourmash signatures/sketches")
     if not sketches:
         # If sketches is empty --> something wrong happened
-        sketch_filenames = '\n'.join(metadata[sig_col].head())
-        raise ValueError(f"Could not load sourmash signatures/sketches from"
-                         f" {metadata_csv}! These are some of the files we couldn't "
-                         f"load:\n---\n{sketch_filenames}\n---\nMaybe the molecule or "
-                         f"ksize is wrong? Molecule: {molecule} and ksize: {ksize}")
+        sketch_filenames = "\n".join(metadata[sig_col].head())
+        raise ValueError(
+            f"Could not load sourmash signatures/sketches from"
+            f" {metadata_csv}! These are some of the files we couldn't "
+            f"load:\n---\n{sketch_filenames}\n---\nMaybe the molecule or "
+            f"ksize is wrong? Molecule: {molecule} and ksize: {ksize}"
+        )
     sketch_series = pd.Series(sketches, index=[x.name() for x in sketches])
     logger.info(f"\nSketch series head: {sketch_series.head()}")
 
     # If group1 is provided, only do one hash enrichment
     if group1 is not None:
         logger.info(f"\n--- group: {group1} ---")
-        coefficients = get_hashes_enriched_in_group(group1, metadata, group_col,
-                                                    sketch_series, verbose=verbose, C=C,
-                                                    n_jobs=n_jobs, solver=solver,
-                                                    penalty=penalty,
-                                                    random_state=random_state,
-                                                    max_group_size=max_group_size,
-                                                    with_abundance=with_abundance)
+        coefficients = get_hashes_enriched_in_group(
+            group1,
+            metadata,
+            group_col,
+            sketch_series,
+            verbose=verbose,
+            C=C,
+            n_jobs=n_jobs,
+            solver=solver,
+            penalty=penalty,
+            random_state=random_state,
+            max_group_size=max_group_size,
+            with_abundance=with_abundance,
+            min_cells=min_cells,
+            min_abundance=min_abundance,
+        )
         write_hash_coefficients(coefficients, group1, threshold)
     else:
         for group1, df in metadata.groupby(group_col):
             logger.info(f"\n--- group: {group1} ---")
-            coefficients = get_hashes_enriched_in_group(group1, metadata, group_col,
-                                                        sketch_series, verbose=verbose,
-                                                        C=C,
-                                                        n_jobs=n_jobs, solver=solver,
-                                                        penalty=penalty,
-                                                        random_state=random_state,
-                                                        max_group_size=max_group_size,
-                                                        with_abundance=with_abundance)
+            coefficients = get_hashes_enriched_in_group(
+                group1,
+                metadata,
+                group_col,
+                sketch_series,
+                verbose=verbose,
+                C=C,
+                n_jobs=n_jobs,
+                solver=solver,
+                penalty=penalty,
+                random_state=random_state,
+                max_group_size=max_group_size,
+                with_abundance=with_abundance,
+                min_cells=min_cells,
+                min_abundance=min_abundance,
+            )
             write_hash_coefficients(coefficients, group1, threshold)
 
 
 def write_hash_coefficients(coefficients, group, threshold):
     # No funny characters, and all lowercase, no spaces
-    sanitized = sanitize_filename(group).lower().replace(' ', '_')
+    sanitized = sanitize_filename(group).lower().replace(" ", "_")
 
     # Write hashes with coefficients to file
-    csv = f'{sanitized}__hash_coefficients.csv'
+    csv = f"{sanitized}__hash_coefficients.csv"
     coefficients.to_csv(csv, header=False)
 
     # Write only hashes above threshold to file
-    filtered_coef = coefficients[coefficients > threshold]
-    txt = f'{sanitized}__informative_hashes.txt'
-    informative_hashes = pd.Series(filtered_coef.index)
-    informative_hashes.to_csv(txt, index=False, header=False)
+    informative_hashes = coefficients.loc[coefficients[COEF_COL] > threshold, group]
+    # Replace all 0s with 1s
+    # informative_hashes = informative_hashes.replace(0, 1)
+    txt = f"{sanitized}__informative_hashes.csv"
+    informative_hashes.to_csv(txt, index=True, header=False)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description="""Perform logistic regression on """)
-    parser.add_argument("--metadata-csv", type=str,
-                        help="CSV of metadata with columns: sample_id,fasta,sig,group")
-    parser.add_argument('-k', '--ksize', type=int, required=True)
+    parser = argparse.ArgumentParser(description="""Perform logistic regression on """)
     parser.add_argument(
-        '--input-is-protein', action='store_true',
-        help='Consume protein sequences - no translation needed.'
+        "--metadata-csv",
+        type=str,
+        help="CSV of metadata with columns: sample_id,fasta,sig,group",
+    )
+    parser.add_argument("-k", "--ksize", type=int, required=True)
+    parser.add_argument(
+        "--input-is-protein",
+        action="store_true",
+        help="Consume protein sequences - no translation needed.",
     )
     parser.add_argument(
-        '--with-abundance', action='store_true',
-        help='Include hash abundances for differential hash expression'
+        "--with-abundance",
+        action="store_true",
+        help="Include hash abundances for differential hash expression",
     )
-    parser.add_argument('-g', "--group-col", type=str,
-                        default='group',
-                        help="Name of column in metadata containing paths to signature "
-                             "files ")
-    parser.add_argument('-g1', "--group1", type=str,
-                        default=None,
-                        help="If provided, only do differential hash enrichment"
-                             " for this group vs the rest")
-    parser.add_argument('-s', "--sig-col", type=str,
-                        default='sig',
-                        help="Name of column in metadata to group by to find "
-                             "differential hash groups")
-    parser.add_argument('-t', "--threshold", type=float, default=0,
-                        help="Value to use to get high-scoring hashes")
-    parser.add_argument("--solver", type=str, default='saga',
-                        help="""From scikit-learn Logistic Regression documentation:
+    parser.add_argument(
+        "-g",
+        "--group-col",
+        type=str,
+        default="group",
+        help="Name of column in metadata containing paths to signature " "files ",
+    )
+    parser.add_argument(
+        "-g1",
+        "--group1",
+        type=str,
+        default=None,
+        help="If provided, only do differential hash enrichment"
+        " for this group vs the rest",
+    )
+    parser.add_argument(
+        "-s",
+        "--sig-col",
+        type=str,
+        default="sig",
+        help="Name of column in metadata to group by to find "
+        "differential hash groups",
+    )
+    parser.add_argument(
+        "-t",
+        "--threshold",
+        type=float,
+        default=0,
+        help="Value to use to get high-scoring hashes",
+    )
+    parser.add_argument(
+        "--solver",
+        type=str,
+        default="saga",
+        help="""From scikit-learn Logistic Regression documentation:
 Algorithm to use in the optimization problem.
 - For small datasets, 'liblinear' is a good choice, whereas 'sag' and 'saga' are faster
   for large ones.
@@ -243,39 +378,77 @@ Algorithm to use in the optimization problem.
 - 'liblinear' does not support setting penalty='none'
 Note that 'sag' and 'saga' fast convergence is only guaranteed on features with
 approximately the same scale. You can preprocess the data with a scaler from
-sklearn.preprocessing.""")
-    parser.add_argument('-C', "--inverse-regularization-strength", type=float,
-                        default=0.1,
-                        help="From scikit-learn Logistic Regression documentation: "
-                             "Inverse of regularization strength; must be a positive "
-                             "float. Like in support vector machines, smaller values "
-                             "specify stronger regularization."
-                             "\n(aka smaller values --> "
-                             "fewer 'informative' features which is easier to follow "
-                             "up on)")
-    parser.add_argument("--penalty", type=str, default=PENALTY,
-                        help="From scikit-learn Logistic Regression documentation: "
-                             "Inverse of "
-                             "regularization strength; must be a positive float. Like "
-                             "in support vector machines, smaller values specify"
-                             " stronger regularization. (aka smaller values --> fewer "
-                             "'informative' features which is easier to follow up on)")
-    parser.add_argument('-p', '--n-jobs', type=int, default=1,
-                        help='Number of concurrent processes to use for'
-                             ' joblib.Parallel')
-    parser.add_argument('-m', '--max-group-size', type=int, default=MAX_GROUP_SIZE,
-                        help='If a group is larger than this, subsample random cells '
-                             '(using the --random-state) ')
-    parser.add_argument('-r', '--random-state', type=int, default=0,
-                        help='Set seed of random number generator to ensure '
-                             'reproducible results')
-    parser.add_argument('-v', "--verbose", action='store_true',
-                        help="If true, have lots of output")
-    parser.add_argument("--use-sig-basename", action='store_true',
-                        help="If true, trim the folder name from the signature path in "
-                             "the metadata csv. Useful primarily for Nextflow "
-                             "pipelines, as the files needed for each process are soft"
-                             " linked into the working folder")
+sklearn.preprocessing.""",
+    )
+    parser.add_argument(
+        "-C",
+        "--inverse-regularization-strength",
+        type=float,
+        default=0.1,
+        help="From scikit-learn Logistic Regression documentation: "
+        "Inverse of regularization strength; must be a positive "
+        "float. Like in support vector machines, smaller values "
+        "specify stronger regularization."
+        "\n(aka smaller values --> "
+        "fewer 'informative' features which is easier to follow "
+        "up on)",
+    )
+    parser.add_argument(
+        "--penalty",
+        type=str,
+        default=PENALTY,
+        help="From scikit-learn Logistic Regression documentation: "
+        "Inverse of "
+        "regularization strength; must be a positive float. Like "
+        "in support vector machines, smaller values specify"
+        " stronger regularization. (aka smaller values --> fewer "
+        "'informative' features which is easier to follow up on)",
+    )
+    parser.add_argument(
+        "-p",
+        "--n-jobs",
+        type=int,
+        default=1,
+        help="Number of concurrent processes to use for" " joblib.Parallel",
+    )
+    parser.add_argument(
+        "-m",
+        "--max-group-size",
+        type=int,
+        default=MAX_GROUP_SIZE,
+        help="If a group is larger than this, subsample random cells "
+        "(using the --random-state) ",
+    )
+    parser.add_argument(
+        "--min-cells",
+        type=int,
+        default=MIN_CELLS,
+        help="Only use hashes expressed in at least this many cells",
+    )
+    parser.add_argument(
+        "--min-abundance",
+        type=int,
+        default=MIN_ABUNDANCE,
+        help="Only use hashes with at least this much abundance",
+    )
+    parser.add_argument(
+        "-r",
+        "--random-state",
+        type=int,
+        default=0,
+        help="Set seed of random number generator to ensure reproducible results",
+    )
+    parser.add_argument(
+        "-v", "--verbose", action="store_true", help="If true, have lots of output"
+    )
+    parser.add_argument(
+        "--use-sig-basename",
+        action="store_true",
+        help="If true, trim the folder name from the signature path in "
+        "the metadata csv. Useful primarily for Nextflow "
+        "pipelines, as the files needed for each process are soft"
+        " linked into the working folder",
+    )
 
     add_construct_moltype_args(parser)
     args = parser.parse_args()
@@ -283,25 +456,29 @@ sklearn.preprocessing.""")
     # Ensure that protein ksizes are divisible by 3
     if (args.protein or args.dayhoff or args.hp) and not args.input_is_protein:
         if args.ksize % 3 != 0:
-            error('protein ksizes must be divisible by 3, sorry!')
-            error('bad ksizes: {}', ", ".join(args.ksize))
+            error("protein ksizes must be divisible by 3, sorry!")
+            error("bad ksizes: {}", ", ".join(args.ksize))
             sys.exit(-1)
 
     moltype = calculate_moltype(args)
 
-    main(metadata_csv=args.metadata_csv,
-         ksize=args.ksize,
-         molecule=moltype,
-         group_col=args.group_col,
-         group1=args.group1,
-         sig_col=args.sig_col,
-         threshold=args.threshold,
-         verbose=args.verbose,
-         C=args.inverse_regularization_strength,
-         solver=args.solver,
-         penalty=args.penalty,
-         n_jobs=args.n_jobs,
-         random_state=args.random_state,
-         use_sig_basename=args.use_sig_basename,
-         max_group_size=args.max_group_size,
-         with_abundance=args.with_abundance)
+    main(
+        metadata_csv=args.metadata_csv,
+        ksize=args.ksize,
+        molecule=moltype,
+        group_col=args.group_col,
+        group1=args.group1,
+        sig_col=args.sig_col,
+        threshold=args.threshold,
+        verbose=args.verbose,
+        C=args.inverse_regularization_strength,
+        solver=args.solver,
+        penalty=args.penalty,
+        n_jobs=args.n_jobs,
+        random_state=args.random_state,
+        use_sig_basename=args.use_sig_basename,
+        max_group_size=args.max_group_size,
+        with_abundance=args.with_abundance,
+        min_cells=args.min_cells,
+        min_abundance=args.min_abundance,
+    )