+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Initial release of nf-core/predictorthologs, created with the [nf-core](http://n
 - Added option of using sourmash to search instead of DIAMOND. This stays in hash-land by searching hashes (provided by `--hashes` or found by `--diff_hash_expression`) directly into a sourmash sequence bloom tree index database
 - Added ability to filter `bam` file for reads containing hashes, if the `--csv` contains a column named `bam` and the option `--filter-bam-hashes` is provided
 - Added option for bam deduplication, if you wish to skip deduplication step add the `-skip_remove_duplicates_bam` flag
-- Added ability to search DIAMOND for hashes that were unassigned from sourmash ([#79](https://github.com/czbiohub/nf-predictorthologs/pull/79))
+- For differential hash expression (when `--diff_hash_expression` is provided), first use `sourmash gather` to search, then extract k-mers for unassigned hashes, then convert hashes to kmers and sequences, and search via DIAMOND for the unassigned hashes ([#79](https://github.com/czbiohub/nf-predictorthologs/pull/79))
 
 ### `Fixed`